storage: Less buttons in the content tabs
They should all be in the menu now, but got left behind.
